requirements.txt
openai
python-dotenv
streamlit
pypdf
chromadb
langchain
langchain-openai
langchain-community
tiktoken
duckduckgo-search
fastapi
uvicorn
ddgs
